Emily's first draft
My first round of edits include a lot of inline comments and questions. You'll find them by searching "<!--EM". I think we're relying too much on the images to tell the reader what to do but we need to explicitly spell out the steps for accessibility and SEO reasons. 

Also, I dramatically simplified the introduction and removed the  Appsody content that was duplicated from other documentation. I offered a streamlined introduction and highlighted prerequisites and steps. If you have a suggestion of how long this tutorial will take after the prerequisites are installed, let me know.